HMW-729 Fixed issue of icons not showing up next to glossary terms rendered after mount.
