+++ app/client/src/components/shared/DataContent.js
@@ -2,10 +2,8 @@
 /** @jsxImportSource @emotion/react */
 
 import { useEffect } from 'react';
-import { createRoot } from 'react-dom/client';
 import { css } from '@emotion/react';
 // components
-import { GlossaryTerm } from 'components/shared/GlossaryPanel';
 import { linkButtonStyles } from 'components/shared/LinkButton';
 import LoadingSpinner from 'components/shared/LoadingSpinner';
 // config
@@ -156,24 +154,6 @@ function DataContent() {
   }, []);
 
   const { data, status } = useDataSourcesContext();
-
-  useEffect(() => {
-    if (status !== 'success') return;
-    const spans = Array.from(document.querySelectorAll('span'));
-    spans.forEach((span) => {
-      if (
-        !span.dataset.hasOwnProperty('glossaryTerm') ||
-        !span.dataset.hasOwnProperty('term')
-      )
-        return;
-
-      const node = document.createElement('span');
-      createRoot(node).render(
-        <GlossaryTerm term={span.dataset.term}>{span.innerText}</GlossaryTerm>,
-      );
-      span.parentNode.replaceChild(node, span);
-    });
-  }, [status]);
 
   if (status === 'fetching') return <LoadingSpinner />;
